# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "description": "FastAPI's HTTPException body; every non-2xx below uses this shape.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "LoginOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Seconds. PI_TOKEN_TTL_MINUTES * 60.",
            "title": "Expires In",
            "type": "integer"
          },
          "is_admin": {
            "description": "So the client can show/hide the admin console right after login.",
            "title": "Is Admin",
            "type": "boolean"
          },
          "token_type": {
            "const": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "username": {
            "description": "The stored account name, not an echo of the request. MySQL's default collation matches case-insensitively, so the login form's spelling and the account's spelling can differ.",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "username",
          "is_admin"
        ],
        "title": "LoginOut",
        "type": "object"
      },
...
      "MemoryClearOut": {
        "properties": {
          "deleted": {
            "description": "How many facts this user had; 0 when memory is off.",
            "title": "Deleted",
            "type": "integer"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "MemoryClearOut",
        "type": "object"
      },
...
      "MemoryListOut": {
        "properties": {
          "facts": {
            "description": "Oldest first, capped at 500.",
            "items": {
              "$ref": "#/components/schemas/MemoryFactOut"
            },
            "title": "Facts",
            "type": "array"
          },
          "status": {
            "description": "'ready', 'disabled' or 'unavailable: <reason>'. 'disabled' with an empty list is the normal state when PI_MILVUS_URI or PI_EMBEDDING_MODEL is unset, not an error.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "facts",
          "status"
        ],
        "title": "MemoryListOut",
        "type": "object"
      },
...
      "RegisterIn": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 32,
            "minLength": 2,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterIn",
        "type": "object"
      },
      "RegisterOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_admin": {
            "description": "Always false. Admin is granted by editing the DB.",
            "title": "Is Admin",
            "type": "boolean"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "is_admin"
        ],
        "title": "RegisterOut",
        "type": "object"
      },
...
      "RunIn": {
        "properties": {
          "builtin_tools": {
            "description": "Gateway-executed tools to offer the model this run.",
            "items": {
              "enum": [
                "web_search",
                "web_extractor",
                "code_interpreter"
              ],
              "type": "string"
            },
            "maxItems": 3,
            "title": "Builtin Tools",
            "type": "array"
          },
          "enable_search": {
            "default": false,
            "description": "Endpoint-side web search woven into the answer (extra_body.enable_search).",
            "title": "Enable Search",
            "type": "boolean"
          },
          "files": {
            "description": "Names of files uploaded to this session to attach to the prompt.",
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Files",
            "type": "array"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt": {
            "maxLength": 32000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "RunIn",
        "type": "object"
      },
      "SessionCreatedOut": {
        "description": "POST /v1/sessions. cwd is echoed so the client can show where bash runs.",
        "properties": {
          "cwd": {
            "title": "Cwd",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "model",
          "cwd"
        ],
        "title": "SessionCreatedOut",
        "type": "object"
      },
      "SessionIn": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "title": {
            "default": "session",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "SessionIn",
        "type": "object"
      },
      "SessionListOut": {
        "properties": {
          "sessions": {
            "description": "Newest first, capped at 50.",
            "items": {
              "$ref": "#/components/schemas/SessionSummary"
            },
            "title": "Sessions",
            "type": "array"
          }
        },
        "required": [
          "sessions"
        ],
        "title": "SessionListOut",
        "type": "object"
      },
...
      "SseCompactionData": {
        "properties": {
          "chars_after": {
            "title": "Chars After",
            "type": "integer"
          },
          "chars_before": {
            "title": "Chars Before",
            "type": "integer"
          },
          "dropped": {
            "description": "Number of messages replaced by the summary.",
            "title": "Dropped",
            "type": "integer"
          }
        },
        "required": [
          "dropped",
          "chars_before",
          "chars_after"
        ],
        "title": "SseCompactionData",
        "type": "object"
      },
...
      "SsePlanData": {
        "description": "Inherits Plan instead of re-declaring title/steps, so the bounds live in one place.",
        "properties": {
          "steps": {
            "description": "Ordered, concrete steps.",
            "items": {
              "maxLength": 300,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          },
          "title": {
            "description": "One-line goal of the task.",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "steps"
        ],
        "title": "SsePlanData",
        "type": "object"
      },
      "SseStartData": {
        "properties": {
          "model": {
            "description": "The model this turn actually resolved to.",
            "title": "Model",
            "type": "string"
          },
          "session": {
            "title": "Session",
            "type": "string"
          }
        },
        "required": [
          "session",
          "model"
        ],
        "title": "SseStartData",
        "type": "object"
      },
      "SseTextDeltaData": {
        "properties": {
          "text": {
            "description": "One incremental chunk of assistant text.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "SseTextDeltaData",
        "type": "object"
      },
      "SseToolCallEndData": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ok": {
            "description": "False when the tool raised, when the policy denied it, or when a terminal tool earlier in the same batch ended the turn and this call was skipped without executing - in that last case `result` says so, and the tool had no side effects.",
            "title": "Ok",
            "type": "boolean"
          },
          "result": {
            "description": "Preview only: AgentLoop cuts it to 200 characters and flattens newlines to spaces before emitting the event. Fetch the full result from GET /v1/sessions/{id}/messages instead.",
            "title": "Result",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "ok",
          "result"
        ],
        "title": "SseToolCallEndData",
        "type": "object"
      },
      "SseToolCallStartData": {
        "properties": {
          "id": {
            "description": "Tool call id; matches the id in toolcall_end. Emitted once per streamed argument chunk, so the same id arrives several times - key on the id, do not treat each frame as a new call.",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "SseToolCallStartData",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "pi-py server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "'invalid credentials' for an unknown username or a wrong password (deliberately indistinguishable), 'account disabled' only when the password was correct - so the second message does not enumerate accounts either."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
...
    "/v1/auth/register": {
      "post": {
        "description": "Open signup, always a normal user; admin is granted by editing the DB.\n\nUnauthenticated by design, so 8300 must not be reachable from the internet.",
        "operationId": "register_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Username does not match ^[a-zA-Z0-9_-]{2,32}$. The password bounds are enforced earlier, by the request body schema (422)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Username already exists - including a signup that lost the race to an identical one."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register"
      }
    },
...
    "/v1/memories": {
      "delete": {
        "description": "Forget everything. This is the erasure endpoint (PIPL art. 47 / GDPR art. 17).\n\nReturns a count rather than a flag so the caller can tell \"you had nothing\nstored\" from \"memory is not enabled here\" without a second request.",
        "operationId": "clear_memories_v1_memories_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryClearOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, malformed, expired or revoked bearer token - or the account was disabled or deleted since the token was issued."
          }
        },
        "summary": "Clear Memories"
      },
      "get": {
        "description": "The caller's stored long-term facts, oldest first.\n\nThe user id comes from the token and never from a parameter, so there is no\nway to ask for somebody else's memory. `limit` is capped server-side at 500.",
        "operationId": "list_memories_v1_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, malformed, expired or revoked bearer token - or the account was disabled or deleted since the token was issued."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Memories"
      }
    },
...
    "/v1/sessions": {
      "get": {
        "operationId": "list_sessions_v1_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, malformed, expired or revoked bearer token - or the account was disabled or deleted since the token was issued."
          }
        },
        "summary": "List Sessions"
      },
      "post": {
        "operationId": "create_session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreatedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, malformed, expired or revoked bearer token - or the account was disabled or deleted since the token was issued."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session"
      }
    },
...
    "/v1/sessions/{session_id}/runs": {
      "post": {
        "operationId": "run_v1_sessions__session_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SseStartData"
                    },
                    {
                      "$ref": "#/components/schemas/SseTextDeltaData"
                    },
                    {
                      "$ref": "#/components/schemas/SseToolCallStartData"
                    },
                    {
                      "$ref": "#/components/schemas/SseToolCallEndData"
                    },
                    {
                      "$ref": "#/components/schemas/SseCompactionData"
                    },
                    {
                      "$ref": "#/components/schemas/SsePlanData"
                    },
                    {
                      "$ref": "#/components/schemas/SseTurnEndData"
                    },
                    {
                      "$ref": "#/components/schemas/SseErrorData"
                    },
                    {
                      "$ref": "#/components/schemas/SseDoneData"
                    }
                  ]
                }
              }
            },
            "description": "Server-Sent Events, over **POST** - so `EventSource` cannot be used. Read\n`response.body` with a `ReadableStream` reader, buffer on blank lines, and take\nthe `event:` / `data:` fields of each frame (a leading `:` is a comment).\n\nOne turn emits `start`, then any number of `text_delta`, `toolcall_start`,\n`toolcall_end` and `compaction`, then at most one `plan`, then `turn_end`, then\n`done`. `error` can arrive at any point **and the HTTP status is still 200**, so a\nclient that only checks `response.ok` will silently swallow failed runs.\n\nWithin a turn the agent loop streams text first, then tool calls, then runs each\ntool in order - so a `text_delta` arriving *after* a `toolcall_end` means a new\nassistant message has begun. There is no message-boundary event; a client that\nrebuilds a transcript has to infer the boundary that way.\n\n`plan` comes after the last `toolcall_end` and before `turn_end`. It is emitted by\n`submit_plan`, a terminal tool, so the run ends right after it - at most one per\nrun. The same plan is persisted and returned as\n`GET /v1/sessions/{id}`'s `plan` field, so painting it from the stream and reading\nit back after a reload agree.\n\nA run cut short that way still emits exactly one `toolcall_end` for every\n`toolcall_start`, including the calls it skipped; those carry `ok: false` and a\n`result` saying they never executed. No call is ever left looking like it is still\nrunning.\n\nThe `data` payload of each frame is the matching `Sse*Data` schema below; the\nevent name is *not* repeated inside the JSON.\n\nForward compatibility: an event type this version of the server does not know\nhow to serialize is emitted as `event: unknown` with `{}`. Clients must ignore\nevent names they do not recognise instead of failing on them, so a newer server\ncan add events without breaking an older browser tab.\n"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, malformed, expired or revoked bearer token - or the account was disabled or deleted since the token was issued."
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Monthly token quota exhausted. The detail carries the used/limit figures; GET /v1/usage returns them as numbers."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "No such session. Ownership is part of the lookup, so another user's session is also 404 and never 403 - the existence of an id is not leaked across accounts."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Per-user fixed-window run limit (PI_RATE_LIMIT_RUNS_PER_MIN) exceeded.",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds to wait, always at least 1.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Run"
      }
    },
...
  }
}